+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Range prefix support for many different commands
 - Buffer-like commands now accept "#" to specify the last used tab
 - Actions "startFollowNewSplit" and "startFollowNewVerSplit" to open links in splits with follow mode
+- Setting "inputfocusalignment" to align input fields consistently on focus from follow mode
 
 ### Changed